+++ sample/SpendingBuilder.hs
@@ -2,12 +2,32 @@
 
 module SpendingBuilder (specs) where
 
-import Plutarch.Context
+import Plutarch.Context (
+    SpendingBuilder,
+    input,
+    mint,
+    output,
+    pubKey,
+    script,
+    tryBuildSpending,
+    withDatum,
+    withRef,
+    withRefIndex,
+    withRefTxId,
+    withSpendingOutRef,
+    withSpendingOutRefId,
+    withSpendingOutRefIdx,
+    withValue,
+ )
+import PlutusLedgerApi.V2 (
+    ScriptContext (scriptContextPurpose),
+    ScriptPurpose (Spending),
+    TxOutRef (..),
+    singleton,
+ )
+import qualified Prettyprinter as P
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (assertFailure, testCase, (@?=))
-
-import PlutusLedgerApi.V1 (ScriptContext (..), ScriptPurpose (..), TxOutRef (..))
-import PlutusLedgerApi.V1.Value (singleton)
 
 someOutRef :: TxOutRef
 someOutRef = TxOutRef "abcdee" 71
@@ -23,16 +43,16 @@ sample =
         , input $
             pubKey "aaccdd"
                 <> withValue (singleton "cc" "hello" 123)
-                <> withOutRef someOutRef
+                <> withRef someOutRef
         , input $
             pubKey "aabb"
                 <> withValue (singleton "cc" "hello" 123)
                 <> withRefIndex 1121
-                <> withTxId "abababcc"
+                <> withRefTxId "abababcc"
         , input $
             pubKey "eeffdd"
                 <> withValue (singleton "cc" "hello" 123)
-                <> withTxId "eeddaa"
+                <> withRefTxId "eeddaa"
         , input $
             pubKey "eeee"
                 <> withValue (singleton "cc" "hello" 123)
@@ -47,37 +67,38 @@ specs =
     testGroup
         "Spending Builder Unit Tests"
         [ testCase "SpendingBuilder should fail when no input validator identifier is given" $
-            case buildSpending sample of
+            case tryBuildSpending mempty sample of
                 Left _ -> pure ()
                 Right _ -> assertFailure "Builder succeed when it should have failed"
         , testCase "Set input validator identifier with TxOutRef" $
-            case buildSpending (sample <> withSpendingOutRef someOutRef) of
-                Left err -> assertFailure $ "Failed with error : " <> err
+            case tryBuildSpending mempty (sample <> withSpendingOutRef someOutRef) of
+                Left err -> assertFailure $ "Failed with error : " <> show (P.pretty err)
                 Right (scriptContextPurpose -> Spending outref) -> outref @?= someOutRef
                 Right _ -> assertFailure "SpendingBuilder built script context that is not spending"
         , testCase "Set input validator identifier with TxOutRefId" $
-            case buildSpending (sample <> withSpendingOutRefId "abababcc") of
-                Left err -> assertFailure $ "Failed with error : " <> err
+            case tryBuildSpending mempty (sample <> withSpendingOutRefId "abababcc") of
+                Left err -> assertFailure $ "Failed with error : " <> show (P.pretty err)
                 Right (scriptContextPurpose -> Spending outref) -> txOutRefId outref @?= "abababcc"
                 Right _ -> assertFailure "SpendingBuilder built script context that is not spending"
         , testCase "Set input validator identifier with TxOutRefIdx" $
-            case buildSpending (sample <> withSpendingOutRefIdx 19) of
-                Left err -> assertFailure $ "Failed with error : " <> err
+            case tryBuildSpending mempty (sample <> withSpendingOutRefIdx 19) of
+                Left err -> assertFailure $ "Failed with error : " <> show (P.pretty err)
                 Right (scriptContextPurpose -> Spending outref) -> txOutRefIdx outref @?= 19
                 Right _ -> assertFailure "SpendingBuilder built script context that is not spending"
         , testCase "Validator identifier should be override-able" $
-            case buildSpending (sample <> withSpendingOutRefIdx 19 <> withSpendingOutRef someOutRef) of
-                Left err -> assertFailure $ "Failed with error : " <> err
+            case tryBuildSpending mempty (sample <> withSpendingOutRefIdx 19 <> withSpendingOutRef someOutRef) of
+                Left err -> assertFailure $ "Failed with error : " <> show (P.pretty err)
                 Right (scriptContextPurpose -> Spending outref) -> outref @?= someOutRef
                 Right _ -> assertFailure "SpendingBuilder built script context that is not spending"
         , testCase "Validator identifier should be override-able 2" $
-            case buildSpending
+            case tryBuildSpending
+                mempty
                 ( sample
                     <> withSpendingOutRefIdx 19
                     <> withSpendingOutRef someOutRef
                     <> withSpendingOutRefId "abababcc"
                 ) of
-                Left err -> assertFailure $ "Failed with error : " <> err
+                Left err -> assertFailure $ "Failed with error : " <> show (P.pretty err)
                 Right (scriptContextPurpose -> Spending outref) -> txOutRefId outref @?= "abababcc"
                 Right _ -> assertFailure "SpendingBuilder built script context that is not spending"
         ]